remove getAllowedKubeUsersAndGroupsForCluster from roleset and move it into server
